Add ES6, node_modules to Asset Path,Pipeline
This is a giant beast. In future, we'll switch off this Asset Pipeline
completely. Meantime, we continuse to use it follow this references:

- https://github.com/reactjs/react-rails/issues/413
- https://github.com/gauravtiwari/browserify_on_rails/

Basically, we use ES6 import on `components.js`. We then expose those
components into a global variable. Then sprocket include them and we use
`react-rails` to render those component.